Removed the commented parts of the files
